// backend/services/auction-service/src/types/event.enums.ts
export enum EventType {
  // AUTH Category (5 events)
  AUTH_LOGIN_SUCCESS = 'AUTH_LOGIN_SUCCESS',
  AUTH_LOGIN_FAILED = 'AUTH_LOGIN_FAILED',
  AUTH_LOGOUT = 'AUTH_LOGOUT',
  AUTH_TOKEN_ISSUED = 'AUTH_TOKEN_ISSUED',
  AUTH_TOKEN_REVOKED = 'AUTH_TOKEN_REVOKED',
  
  // SEARCH Category (4 events)
  SEARCH_QUERY_EXECUTED = 'SEARCH_QUERY_EXECUTED',
  SEARCH_FILTER_APPLIED = 'SEARCH_FILTER_APPLIED',
  SEARCH_RESULT_VIEWED = 'SEARCH_RESULT_VIEWED',
  SEARCH_RECOMMENDATION_SHOWN = 'SEARCH_RECOMMENDATION_SHOWN',
  
  // PRODUCT Category (6 events)
  PRODUCT_CREATED = 'PRODUCT_CREATED',
  PRODUCT_UPDATED = 'PRODUCT_UPDATED',
  PRODUCT_PUBLISHED = 'PRODUCT_PUBLISHED',
  PRODUCT_UNPUBLISHED = 'PRODUCT_UNPUBLISHED',
  PRODUCT_DELETED = 'PRODUCT_DELETED',
  PRODUCT_VIEWED = 'PRODUCT_VIEWED',
  
  // AUCTION Category (8 events)
  AUCTION_CREATED = 'AUCTION_CREATED',
  AUCTION_STARTED = 'AUCTION_STARTED',
  AUCTION_ENDED_NORMAL = 'AUCTION_ENDED_NORMAL',
  AUCTION_ENDED_RESERVE_NOT_MET = 'AUCTION_ENDED_RESERVE_NOT_MET',
  AUCTION_EXTENDED = 'AUCTION_EXTENDED',
  AUCTION_CANCELLED = 'AUCTION_CANCELLED',
  AUCTION_SETTLED = 'AUCTION_SETTLED',
  AUCTION_FINALIZED = 'AUCTION_FINALIZED',
  
  // BID Category (7 events)
  BID_PLACED = 'BID_PLACED',
  BID_OUTBID = 'BID_OUTBID',
  BID_WON = 'BID_WON',
  BID_CANCELLED = 'BID_CANCELLED',
  BID_INVALIDATED = 'BID_INVALIDATED',
  BID_THROTTLED = 'BID_THROTTLED',
  PROXY_BID_ACTIVATED = 'PROXY_BID_ACTIVATED',
  
  // ESCROW Category (5 events)
  ESCROW_CREATED = 'ESCROW_CREATED',
  ESCROW_HELD = 'ESCROW_HELD',
  ESCROW_RELEASED = 'ESCROW_RELEASED',
  ESCROW_REFUNDED = 'ESCROW_REFUNDED',
  ESCROW_DISPUTE_FLAGGED = 'ESCROW_DISPUTE_FLAGGED',
  
  // WALLET Category (5 events)
  WALLET_CREATED = 'WALLET_CREATED',
  WALLET_BALANCE_VIEWED = 'WALLET_BALANCE_VIEWED',
  WALLET_TRANSACTION_VIEWED = 'WALLET_TRANSACTION_VIEWED',
  WALLET_TRANSFER_INITIATED = 'WALLET_TRANSFER_INITIATED',
  WALLET_TRANSFER_COMPLETED = 'WALLET_TRANSFER_COMPLETED',
  
  // PAYMENT Category (6 events)
  PAYMENT_INITIATED = 'PAYMENT_INITIATED',
  PAYMENT_INTENT_CREATED = 'PAYMENT_INTENT_CREATED',
  PAYMENT_COMPLETED = 'PAYMENT_COMPLETED',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  PAYMENT_REFUNDED = 'PAYMENT_REFUNDED',
  PAYMENT_WEBHOOK_RECEIVED = 'PAYMENT_WEBHOOK_RECEIVED',
  
  // DELIVERY Category (6 events)
  DELIVERY_CREATED = 'DELIVERY_CREATED',
  DELIVERY_PICKED_UP = 'DELIVERY_PICKED_UP',
  DELIVERY_IN_TRANSIT = 'DELIVERY_IN_TRANSIT',
  DELIVERY_DELIVERED = 'DELIVERY_DELIVERED',
  DELIVERY_FAILED = 'DELIVERY_FAILED',
  DELIVERY_CANCELLED = 'DELIVERY_CANCELLED',
  
  // DISPUTE Category (6 events)
  DISPUTE_CREATED = 'DISPUTE_CREATED',
  DISPUTE_EVIDENCE_SUBMITTED = 'DISPUTE_EVIDENCE_SUBMITTED',
  DISPUTE_UNDER_REVIEW = 'DISPUTE_UNDER_REVIEW',
  DISPUTE_RESOLVED = 'DISPUTE_RESOLVED',
  DISPUTE_ESCALATED = 'DISPUTE_ESCALATED',
  DISPUTE_APPEALED = 'DISPUTE_APPEALED',
  
  // TRUST Category (8 events)
  TRUST_SCORE_CALCULATED = 'TRUST_SCORE_CALCULATED',
  TRUST_SCORE_UPDATED = 'TRUST_SCORE_UPDATED',
  TRUST_ACTION_CREATED = 'TRUST_ACTION_CREATED',
  TRUST_ACTION_LIFTED = 'TRUST_ACTION_LIFTED',
  TRUST_ACTION_EXPIRED = 'TRUST_ACTION_EXPIRED',
  TRUST_APPEAL_SUBMITTED = 'TRUST_APPEAL_SUBMITTED',
  TRUST_APPEAL_APPROVED = 'TRUST_APPEAL_APPROVED',
  TRUST_APPEAL_REJECTED = 'TRUST_APPEAL_REJECTED',
  
  // SYSTEM Category (6 events)
  SYSTEM_STARTUP = 'SYSTEM_STARTUP',
  SYSTEM_SHUTDOWN = 'SYSTEM_SHUTDOWN',
  SYSTEM_ERROR = 'SYSTEM_ERROR',
  SYSTEM_WARNING = 'SYSTEM_WARNING',
  SYSTEM_MAINTENANCE_START = 'SYSTEM_MAINTENANCE_START',
  SYSTEM_MAINTENANCE_END = 'SYSTEM_MAINTENANCE_END',
}